use same helper for clustermodifier as in other tools
